+++ spec/simmer_spec.rb
@@ -14,18 +14,20 @@ describe Simmer do
   let(:src_config_path) { File.join('spec', 'config', 'simmer.yaml') }
   let(:config)          { yaml_read(src_config_path) }
 
-  def stage_simmer_config(spoon_dir, args)
+  def stage_simmer_config(spoon_dir, args, timeout_in_seconds = nil)
     dest_config_dir  = File.join('tmp')
     dest_config_path = File.join(dest_config_dir, 'simmer.yaml')
 
     FileUtils.rm(dest_config_path) if File.exist?(dest_config_path)
     FileUtils.mkdir_p(dest_config_dir)
 
+    timeout = timeout_in_seconds || config.dig('spoon_client', 'timeout_in_seconds')
+
     new_config = config.merge(
       'spoon_client' => {
         'dir' => spoon_dir,
         'args' => args,
-        'timeout_in_seconds' => config.dig('spoon_client', 'timeout_in_seconds')
+        'timeout_in_seconds' => timeout
       }
     )
 
@@ -52,7 +54,7 @@ describe Simmer do
           simmer_dir: simmer_dir
         )
 
-        expect(results.pass?).to be false
+        expect(results).not_to be_passing
       end
     end
 
@@ -68,7 +70,7 @@ describe Simmer do
           simmer_dir: simmer_dir
         )
 
-        expect(results.pass?).to be false
+        expect(results).not_to be_passing
       end
     end
 
@@ -84,7 +86,7 @@ describe Simmer do
           simmer_dir: simmer_dir
         )
 
-        expect(results.pass?).to be true
+        expect(results).to be_passing
       end
     end
 
@@ -100,7 +102,35 @@ describe Simmer do
           simmer_dir: simmer_dir
         )
 
-        expect(results.pass?).to be false
+        expect(results).not_to be_passing
+      end
+    end
+
+    context 'when pdi times out' do
+      let(:spoon_path)  { File.join('spec', 'mocks', 'spoon') }
+      let(:args)        { [0, 10] }
+      let(:config_path) { stage_simmer_config(spoon_path, args, 1) }
+
+      it 'fails' do
+        results = described_class.run(
+          spec_path,
+          config_path: config_path,
+          out: out,
+          simmer_dir: simmer_dir
+        )
+
+        expect(results).not_to be_passing
+      end
+
+      it 'records error' do
+        results = described_class.run(
+          spec_path,
+          config_path: config_path,
+          out: out,
+          simmer_dir: simmer_dir
+        )
+
+        expect(results.runner_results.first.errors).to include(/execution expired/)
       end
     end
 
@@ -123,6 +153,40 @@ describe Simmer do
         expected = config.dig('spoon_client', 'timeout_in_seconds')
 
         expect(actual).to eq(expected)
+      end
+    end
+  end
+
+  context 'when the fixtures are missing' do
+    let(:spec_path)   { File.join('spec', 'fixtures', 'specifications', 'missing_fixtures.yaml') }
+    let(:simmer_dir)  { File.join('spec', 'simmer_spec') }
+    let(:out)         { Out.new }
+    let(:config_path) { stage_simmer_config(spoon_path, args) }
+
+    context 'when pdi does not do anything but does not fail' do
+      let(:spoon_path)  { File.join('spec', 'mocks', 'spoon') }
+      let(:args)        { 0 }
+
+      it 'fails' do
+        results = described_class.run(
+          spec_path,
+          config_path: config_path,
+          out: out,
+          simmer_dir: simmer_dir
+        )
+
+        expect(results).not_to be_passing
+      end
+
+      it 'records error' do
+        results = described_class.run(
+          spec_path,
+          config_path: config_path,
+          out: out,
+          simmer_dir: simmer_dir
+        )
+
+        expect(results.runner_results.first.errors).to include(/fixture missing/)
       end
     end
   end