+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.1.0 (May 12th, 2020)
+
+Additions:
+
+* Do not make missing fixtures short-circuit the rest of the test suite.
+* Do not make PDI timeouts short-circuit the rest of the test suite.
+* Report PDI's exit code and execution time to the console.
+
 # 2.0.0 (May 11th, 2020)
 
 Breaking Changes: